Switch float to double in graph1 example

+++ examples/graph1.cpp
@@ -8,9 +8,9 @@ int main()
     // Set up a morph::Visual 'scene environment'.
     morph::Visual v(1024, 768, "Made with morph::GraphVisual");
     // Create a new GraphVisual with offset within the scene of 0,0,0
-    auto gv = new morph::GraphVisual<float> (v.shaderprog, v.tshaderprog, {0,0,0});
+    auto gv = new morph::GraphVisual<double> (v.shaderprog, v.tshaderprog, {0,0,0});
     // Data for the x axis. A vVector is like std::vector, but with built-in maths methods
-    morph::vVector<float> x;
+    morph::vVector<double> x;
     // This works like numpy's linspace() (the 3 args are "start", "end" and "num"):
     x.linspace (-0.5, 0.8, 14);
     // Set a graph up of y = x^3
